Moved Teknik pastebin and url shortener services to separate folders.
