+++ ShareX.UploadersLib/FileUploaders/Teknik.cs
@@ -39,50 +39,6 @@ namespace ShareX.UploadersLib.FileUploaders
         public override TabPage GetUploadersConfigTabPage(UploadersConfigForm form) => form.tpTeknik;
     }
 
-    public class TeknikTextUploaderService : TextUploaderService
-    {
-        public override TextDestination EnumValue { get; } = TextDestination.Teknik;
-
-        public override Icon ServiceIcon => Resources.Teknik;
-
-        public override bool CheckConfig(UploadersConfig config)
-        {
-            return !string.IsNullOrEmpty(config.TeknikPasteAPIUrl) && !string.IsNullOrEmpty(config.TeknikAuthUrl);
-        }
-
-        public override GenericUploader CreateUploader(UploadersConfig config, TaskReferenceHelper taskInfo)
-        {
-            return new TeknikPaster(config.TeknikOAuth2Info, config.TeknikAuthUrl)
-            {
-                APIUrl = config.TeknikPasteAPIUrl
-            };
-        }
-
-        public override TabPage GetUploadersConfigTabPage(UploadersConfigForm form) => form.tpGist;
-    }
-
-    public class TeknikUrlShortenerService : URLShortenerService
-    {
-        public override UrlShortenerType EnumValue { get; } = UrlShortenerType.Teknik;
-
-        public override Icon ServiceIcon => Resources.Teknik;
-
-        public override bool CheckConfig(UploadersConfig config)
-        {
-            return !string.IsNullOrEmpty(config.TeknikUrlShortenerAPIUrl) && !string.IsNullOrEmpty(config.TeknikAuthUrl);
-        }
-
-        public override URLShortener CreateShortener(UploadersConfig config, TaskReferenceHelper taskInfo)
-        {
-            return new TeknikUrlShortener(config.TeknikOAuth2Info, config.TeknikAuthUrl)
-            {
-                APIUrl = config.TeknikUrlShortenerAPIUrl
-            };
-        }
-
-        public override TabPage GetUploadersConfigTabPage(UploadersConfigForm form) => form.tpGist;
-    }
-
     public enum TeknikExpirationUnit
     {
         Never = 0,
@@ -214,90 +170,6 @@ namespace ShareX.UploadersLib.FileUploaders
         }
     }
 
-    public sealed class TeknikPaster : TextUploader, IOAuth2Basic
-    {
-        private Teknik _Teknik;
-
-        public OAuth2Info AuthInfo { get; set; }
-
-        public string APIUrl { get; set; }
-
-        public TeknikPaster(OAuth2Info oauth, string authUrl)
-        {
-            _Teknik = new Teknik(oauth, authUrl);
-            AuthInfo = _Teknik.AuthInfo;
-        }
-
-        public bool GetAccessToken(string code)
-        {
-            return _Teknik.GetAccessToken(code);
-        }
-
-        public string GetAuthorizationURL()
-        {
-            return _Teknik.GetAuthorizationURL();
-        }
-
-        public override UploadResult UploadText(string text, string fileName)
-        {
-            Dictionary<string, string> args = new Dictionary<string, string>();
-            args.Add("code", text);
-
-            string response = SendRequestMultiPart(APIUrl, args, _Teknik.GetAuthHeaders(), null, HttpMethod.POST);
-            TeknikPasteResponseWrapper apiResponse = JsonConvert.DeserializeObject<TeknikPasteResponseWrapper>(response);
-            
-            UploadResult ur = new UploadResult();
-            if (apiResponse.Result != null && apiResponse.Error == null)
-            {
-                ur.URL = apiResponse.Result.Url;
-            }
-
-            return ur;
-        }
-    }
-
-    public sealed class TeknikUrlShortener : URLShortener, IOAuth2Basic
-    {
-        private Teknik _Teknik;
-
-        public OAuth2Info AuthInfo { get; set; }
-
-        public string APIUrl { get; set; }
-
-        public TeknikUrlShortener(OAuth2Info oauth, string authUrl)
-        {
-            _Teknik = new Teknik(oauth, authUrl);
-            AuthInfo = _Teknik.AuthInfo;
-        }
-
-        public bool GetAccessToken(string code)
-        {
-            return _Teknik.GetAccessToken(code);
-        }
-
-        public string GetAuthorizationURL()
-        {
-            return _Teknik.GetAuthorizationURL();
-        }
-
-        public override UploadResult ShortenURL(string url)
-        {
-            Dictionary<string, string> args = new Dictionary<string, string>();
-            args.Add("url", url);
-
-            string response = SendRequestMultiPart(APIUrl, args, _Teknik.GetAuthHeaders(), null, HttpMethod.POST);
-            TeknikUrlShortenerResponseWrapper apiResponse = JsonConvert.DeserializeObject<TeknikUrlShortenerResponseWrapper>(response);
-
-            UploadResult ur = new UploadResult();
-            if (apiResponse.Result != null && apiResponse.Error == null)
-            {
-                ur.ShortenedURL = apiResponse.Result.shortUrl;
-            }
-
-            return ur;
-        }
-    }
-
     public class TeknikErrorResponse
     {
         public string Message { get; set; }
@@ -320,32 +192,5 @@ namespace ShareX.UploadersLib.FileUploaders
         public string IV { get; set; }
         public int BlockSize { get; set; }
         public string DeletionKey { get; set; }
-    }
-
-    public class TeknikPasteResponseWrapper
-    {
-        public TeknikPasteResponse Result { get; set; }
-        public TeknikErrorResponse Error { get; set; }
-    }
-
-    public class TeknikPasteResponse
-    {
-        public string Id { get; set; }
-        public string Url { get; set; }
-        public string Title { get; set; }
-        public string Syntax { get; set; }
-        public string Password { get; set; }
-    }
-
-    public class TeknikUrlShortenerResponseWrapper
-    {
-        public TeknikUrlShortenerResponse Result { get; set; }
-        public TeknikErrorResponse Error { get; set; }
-    }
-
-    public class TeknikUrlShortenerResponse
-    {
-        public string shortUrl { get; set; }
-        public string originalUrl { get; set; }
     }
 }
